Address PR suggestions
