+++ src/Humanizer.Tests/ApiApprover/PublicApiApprovalTest.approve_public_api.approved.txt
@@ -1953,6 +1953,11 @@ namespace Humanizer.Localisation.Ordinalizers
 }
 namespace Humanizer.Localisation.TimeToClockNotation
 {
+    public class BrazilianPortugueseTimeOnlyToClockNotationConverter : Humanizer.Localisation.TimeToClockNotation.ITimeOnlyToClockNotationConverter
+    {
+        public BrazilianPortugueseTimeOnlyToClockNotationConverter() { }
+        public virtual string Convert(System.TimeOnly time) { }
+    }
     public class DefaultTimeOnlyToClockNotationConverter : Humanizer.Localisation.TimeToClockNotation.ITimeOnlyToClockNotationConverter
     {
         public DefaultTimeOnlyToClockNotationConverter() { }
@@ -1961,10 +1966,5 @@ namespace Humanizer.Localisation.TimeToClockNotation
     public interface ITimeOnlyToClockNotationConverter
     {
         string Convert(System.TimeOnly time);
-    }
-    public class PtBrTimeOnlyToClockNotationConverter : Humanizer.Localisation.TimeToClockNotation.ITimeOnlyToClockNotationConverter
-    {
-        public PtBrTimeOnlyToClockNotationConverter() { }
-        public virtual string Convert(System.TimeOnly time) { }
     }
 }
