+++ src/Humanizer/Configuration/TimeOnlyToClockNotationConvertersRegistry.cs
@@ -11,7 +11,7 @@ namespace Humanizer.Configuration
             Register("en-US", new DefaultTimeOnlyToClockNotationConverter());
             Register("en-UK", new DefaultTimeOnlyToClockNotationConverter());
             Register("de",    new DefaultTimeOnlyToClockNotationConverter());
-            Register("pt-BR", new PtBrTimeOnlyToClockNotationConverter());
+            Register("pt-BR", new BrazilianPortugueseTimeOnlyToClockNotationConverter());
         }
     }
 }
